segmentation_system/requirements.txt
Django==4.1.7
django-cleanup==7.0.0
django-import-export==3.1.0
django-simpleui==2023.3.1
matplotlib
numpy
opencv-python